+++ pydantic_ai_harness/environments/local.py
@@ -231,7 +231,7 @@ class LocalEnvironment(AbstractEnvironment):
 
         return results
 
-    async def shell_command(self, command: str, timeout: int | None = None) -> ShellCommandResult:
+    async def shell_command(self, command: str, timeout: float | None = None) -> ShellCommandResult:
         """Run `command` in a shell and return its captured output and exit code.
 
         The command is shell-interpreted (pipes, `&&`, globs, `$VARS` all work) and runs in a fresh
